+++ sdk/src/main/java/com/chainverse/sdk/model/NFT/Listing.java
@@ -3,18 +3,18 @@ package com.chainverse.sdk.model.NFT;
 import java.math.BigInteger;
 
 public class Listing extends BaseInfo {
-    private BigInteger price;
+    private double price;
 
-    public Listing(boolean isEnded, String nft, String owner, String currency, BigInteger tokenId, int fee, BigInteger id, BigInteger price) {
+    public Listing(boolean isEnded, String nft, String owner, String currency, BigInteger tokenId, int fee, BigInteger id, double price) {
         super(isEnded, nft, owner, currency, tokenId, fee, id);
         this.price = price;
     }
 
-    public BigInteger getPrice() {
+    public double getPrice() {
         return price;
     }
 
-    public void setPrice(BigInteger price) {
+    public void setPrice(double price) {
         this.price = price;
     }
 }
